horror-game
This is the code for the horror game. You can edit this code as much as you want. You can you can put this on your computer and save it as an html file, you can put this game in your website ( this game also work on the new google sites), and a lot more.